Add transcript information to a data frame with mutation records

Add transcript information to a data frame with mutation records

AddTranscript(df, trans.ranges = NULL, ref.genome, name.of.VCF = NULL)

Arguments

df

A data frame storing mutation records of a VCF file.

trans.ranges

A data.table which contains transcript range and strand information. Please refer to TranscriptRanges for more details.

ref.genome

A ref.genome argument as described in ICAMS.

name.of.VCF

Name of the VCF file.

Value

A data frame with new columns added to the input data frame, which contain the mutated gene's name, range and strand information.