Standard order of row names in a catalog

This data is designed for those who need to create their own catalogs from formats not supported by this package. The rownames denote the mutation types. For example, for SBS96 catalogs, the rowname AGAT represents a mutation from AGA > ATA.

catalog.row.order

Format

A list of character vectors indicating the standard orders of row names in catalogs.

An object of class list of length 9.

Note

In ID (small insertions and deletions) catalogs, deletion repeat sizes range from 0 to 5+, but for plotting and end-user documentation deletion repeat sizes range from 1 to 6+. In ID83 catalogs, deletion repeat sizes range from 0 to 5.

ID classification

See https://github.com/steverozen/ICAMS/blob/v3.0.9-branch/data-raw/PCAWG7_indel_classification_2021_09_03.xlsx for additional information on ID (small insertions and deletions) mutation classification.

See the documentation for Canonicalize1Del which first handles deletions in homopolymers, then handles deletions in simple repeats with longer repeat units, (e.g. CACACACA, see FindMaxRepeatDel), and if the deletion is not in a simple repeat, looks for microhomology (see FindDelMH).

See the code for unexported function CanonicalizeID and the functions it calls for handling of insertions.

Examples

catalog.row.order$SBS96
#>  [1] "ACAA" "ACCA" "ACGA" "ACTA" "CCAA" "CCCA" "CCGA" "CCTA" "GCAA" "GCCA"
#> [11] "GCGA" "GCTA" "TCAA" "TCCA" "TCGA" "TCTA" "ACAG" "ACCG" "ACGG" "ACTG"
#> [21] "CCAG" "CCCG" "CCGG" "CCTG" "GCAG" "GCCG" "GCGG" "GCTG" "TCAG" "TCCG"
#> [31] "TCGG" "TCTG" "ACAT" "ACCT" "ACGT" "ACTT" "CCAT" "CCCT" "CCGT" "CCTT"
#> [41] "GCAT" "GCCT" "GCGT" "GCTT" "TCAT" "TCCT" "TCGT" "TCTT" "ATAA" "ATCA"
#> [51] "ATGA" "ATTA" "CTAA" "CTCA" "CTGA" "CTTA" "GTAA" "GTCA" "GTGA" "GTTA"
#> [61] "TTAA" "TTCA" "TTGA" "TTTA" "ATAC" "ATCC" "ATGC" "ATTC" "CTAC" "CTCC"
#> [71] "CTGC" "CTTC" "GTAC" "GTCC" "GTGC" "GTTC" "TTAC" "TTCC" "TTGC" "TTTC"
#> [81] "ATAG" "ATCG" "ATGG" "ATTG" "CTAG" "CTCG" "CTGG" "CTTG" "GTAG" "GTCG"
#> [91] "GTGG" "GTTG" "TTAG" "TTCG" "TTGG" "TTTG"
# "ACAA" "ACCA" "ACGA" "ACTA" "CCAA" "CCCA" "CCGA" "CCTA" ...
# There are altogether 96 row names to denote the mutation types
# in SBS96 catalog.