Read in the data lines of a Variant Call Format (VCF) file created by Mutect

Read in the data lines of a Variant Call Format (VCF) file created by Mutect

ReadMutectVCF(file, name.of.VCF = NULL, tumor.col.name = NA)

Arguments

file

The name/path of the VCF file, or a complete URL.

name.of.VCF

Name of the VCF file. If NULL(default), this function will remove all of the path up to and including the last path separator (if any) in file and file path without extensions (and the leading dot) will be used as the name of the VCF file.

tumor.col.name

Name or index of the column in VCF which contains the tumor sample information. It must have quotation marks if specifying the column name. If tumor.col.name is equal to NA(default), this function will use the 10th column to calculate VAFs. See GetMutectVAF for more details.

Value

A data frame storing data lines of a VCF file with two additional columns added which contain the VAF(variant allele frequency) and read depth information.