Plot the a SignatureAnalyzer COMPOSITE signature or catalog into separate pdfs

Plot the a SignatureAnalyzer COMPOSITE signature or catalog into separate pdfs

TestPlotCatCOMPOSITE(catalog, filename.header, type, id = colnames(catalog))

Arguments

catalog

Catalog or signature matrix

filename.header

Contain path and the beginning part of the file name. The name of the pdf files will be: filename.header.SBS.96.pdf filename.header.SBS.1536.pdf filename.header.DBS.78.pdf filename.header.ID.83.pdf

type

See PlotCatalogToPdf.

id

A vector containing the identifiers of the samples or signatures in catalog.