Transcript ranges data

Transcript ranges and strand information for a particular reference genome.

trans.ranges.GRCh37

trans.ranges.GRCh38

trans.ranges.GRCm38

Format

A data.table which contains transcript range and strand information for a particular reference genome. colnames are chrom, start, end, strand, Ensembl.gene.ID, gene.symbol. It uses one-based coordinates.

An object of class data.table (inherits from data.frame) with 19083 rows and 6 columns.

An object of class data.table (inherits from data.frame) with 19096 rows and 6 columns.

An object of class data.table (inherits from data.frame) with 20325 rows and 6 columns.

Source

ftp://ftp.ebi.ac.uk/pub/databases/gencode/Gencode_human/release_30/GRCh37_mapping/gencode.v30lift37.annotation.gff3.gz

ftp://ftp.ebi.ac.uk/pub/databases/gencode/Gencode_human/release_30/gencode.v30.annotation.gff3.gz

ftp://ftp.ebi.ac.uk/pub/databases/gencode/Gencode_mouse/release_M21/gencode.vM21.annotation.gff3.gz

Details

This information is needed to generate catalogs that depend on transcriptional strand information, for example catalogs of class SBS192Catalog.

trans.ranges.GRCh37: Human GRCh37.

trans.ranges.GRCh38: Human GRCh38.

trans.ranges.GRCm38: Mouse GRCm38.

For these two tables, only genes that are associated with a CCDS ID are kept for transcriptional strand bias analysis.

This information is needed for StrelkaSBSVCFFilesToCatalog,
StrelkaSBSVCFFilesToCatalogAndPlotToPdf, MutectVCFFilesToCatalog,
MutectVCFFilesToCatalogAndPlotToPdf, VCFsToSBSCatalogs and VCFsToDBSCatalogs.

Examples

trans.ranges.GRCh37
#> Key: <chrom, start, end>
#>        chrom    start      end strand Ensembl.gene.ID gene.symbol
#>        <ord>    <int>    <int> <char>          <char>      <char>
#>     1:     1    65419    71585      + ENSG00000186092       OR4F5
#>     2:     1   367640   368634      + ENSG00000235249      OR4F29
#>     3:     1   621059   622053      - ENSG00000284662      OR4F16
#>     4:     1   859308   879961      + ENSG00000187634      SAMD11
#>     5:     1   879583   894689      - ENSG00000188976       NOC2L
#>    ---                                                           
#> 19079:     Y 27177048 27208695      - ENSG00000185894       BPY2C
#> 19080:     Y 27768309 27771049      + ENSG00000172288        CDY1
#> 19081:     Y 59100480 59115127      + ENSG00000168939       SPRY3
#> 19082:     Y 59214014 59276439      + ENSG00000124333       VAMP7
#> 19083:     Y 59330252 59343488      + ENSG00000124334        IL9R
# chrom    start      end strand Ensembl.gene.ID  gene.symbol
#     1    65419    71585      + ENSG00000186092        OR4F5
#     1   367640   368634      + ENSG00000235249       OR4F29
#     1   621059   622053      - ENSG00000284662       OR4F16
#     1   859308   879961      + ENSG00000187634       SAMD11
#     1   879583   894689      - ENSG00000188976        NOC2L
#   ...      ...      ...    ...             ...          ...