K-mer abundances

An R list with one element each for BSgenome.Hsapiens.1000genomes.hs37d5,
BSgenome.Hsapiens.UCSC.hg38 and BSgenome.Mmusculus.UCSC.mm10. Each element is in turn a sub-list keyed by exome, transcript, and genome. Each element of the sub list is keyed by the number of rows in the catalog class (as a string, e.g. "78", not 78). The keys are: 78 (DBS78Catalog), 96 (SBS96Catalog), 136 (DBS136Catalog), 144 (DBS144Catalog), 192 (SBS192Catalog), and 1536 (SBS1536Catalog). So, for example to get the exome abundances for SBS96 catalogs for BSgenome.Hsapiens.UCSC.hg38 exomes one would reference
all.abundance[["BSgenome.Hsapiens.UCSC.hg38"]][["exome"]][["96"]]
or all.abundance$BSgenome.Hsapiens.UCSC.hg38$exome$"96". The value of the abundance is an integer vector with the K-mers as names and each value being the count of that K-mer.

all.abundance

Format

See Description.

Examples

all.abundance$BSgenome.Hsapiens.UCSC.hg38$transcript$`144` 
#>        AA        AC        AG        AT        CA        CC        CG        CT 
#>  90769160  57156295  85738416  87552737  83479655  63267896  13730515  87172945 
#>        GA        GC        GG        GT        TA        TC        TG        TT 
#>  71725583  55297347  65778797  65991950  75465411  72188227  93256356 104937932 
#        AA        AC        AG        AT        CA        CC ... 
#  90769160  57156295  85738416  87552737  83479655  63267896 ...
# There are 90769160 AAs on the sense strands of transcripts in
# this genome.